Favourite button and icon added and local storage function added local storage
Added favourite button and heart icon. When favourite button is pressed the heart icon toggles on/ off. A favourited recipe ID is saved to local storage, if the recipe card is opened that has been favourited before, this will be seen in local storage and the heart icon will fill toggle on.